refactor(sandbox-mock): split mock responsibilities
